fix(GameType): change from_id (i32 -> i8)

+++ src/packets/types/game_type.rs
@@ -14,7 +14,7 @@ pub enum GameType {
 impl GameType {
     /// Get the game type from the id
     /// Used to parse it from the TCP stream
-    pub fn from_id(id: i32) -> Option<Self> {
+    pub fn from_id(id: i8) -> Option<Self> {
         match id {
             -1 => Some(Self::NotSet),
             0 => Some(Self::Survival),
